requirements.txt
redis
streamlit
pyyaml